chore: rename `core` directory to `modules`
