+++ src/worker.ts
@@ -1,6 +1,6 @@
 import { platform } from 'angular2/core';
 import { WORKER_APP_APPLICATION, WORKER_APP_PLATFORM } from 'angular2/platform/worker_app';
-import { TASK_PROVIDERS } from './core/task/providers';
+import { TASK_PROVIDERS } from './modules/task/providers';
 import { App } from './components/app/app';
